Update 02intro.asciidoc
While following the book to create a Test Faucet, one of the problems I encountered was that msg.sender is no longer a payable address, and the transfer function cannot be called. Hence, I figured out that payable(msg.sender).transfer(withdraw_amount) is the new method of transferring Ether from a contract, which fixed the issue.